remove foreign key
