--- mlflow/store/db_migrations/versions/7f2a7d5fae7d_add_dataset_tables.py
@@ -1,48 +1,0 @@
-"""add datasets inputs input_tags tables
-
-Revision ID: 7f2a7d5fae7d
-Revises: 3500859a5d39
-Create Date: 2023-03-23 09:48:27.775166
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from mlflow.store.tracking.dbmodels.models import SqlDataset
-
-# revision identifiers, used by Alembic.
-revision = "7f2a7d5fae7d"
-down_revision = "3500859a5d39"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    op.create_table(
-        SqlDataset.__tablename__,
-        sa.Column("dataset_uuid", sa.String(length=36), nullable=False),
-        sa.Column(
-            "experiment_id",
-            sa.Integer(),
-            sa.ForeignKey("experiments.experiment_id"),
-            primary_key=True,
-            nullable=False,
-        ),
-        sa.Column("name", sa.String(length=500), primary_key=True, nullable=False),
-        sa.Column("digest", sa.String(length=36), primary_key=True, nullable=False),
-        sa.Column("dataset_source_type", sa.String(length=36), nullable=False),
-        sa.Column("dataset_source", sa.String(length=5000), nullable=False),
-        sa.Column("dataset_schema", sa.String(length=5000), nullable=True),
-        sa.Column("dataset_profile", sa.String(length=5000), nullable=True),
-        sa.PrimaryKeyConstraint("experiment_id", "name", "digest", name="dataset_pk"),
-        sa.Index(f"index_{SqlDataset.__tablename__}_dataset_uuid", "dataset_uuid", unique=False),
-        sa.Index(
-            f"index_{SqlDataset.__tablename__}_dataset_uuid_dataset_source_type",
-            "experiment_id",
-            "dataset_source_type",
-            unique=False,
-        ),
-    )
-
-
-def downgrade():
-    pass
